requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
itsdangerous>=2.2.0
